% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{add_axis_title_spacing_plotly}
\alias{add_axis_title_spacing_plotly}
\title{Add Axis Spacing for Faceted Plotly}
\usage{
add_axis_title_spacing_plotly(plotly_obj, axis, n_lines)
}
\arguments{
\item{plotly_obj}{An object generated by ggplotly}

\item{axis}{String value of "x" or "y". Which axis title to add spacing to.
If need to add spacing to both, make separate calls to \code{add_axis_title_spacing_plotly()}}

\item{n_lines}{Numeric value. Number of lines worth of spacing to add.}
}
\value{
A the original plotly object with updated axis name
}
\description{
Note: This does not work as expected in Rmarkdown
Add lines of whitespace to x-axis or y-axis title to shift title down or left.
}
\examples{
library(ggplot2)
library(plotly)
library(dplyr)
set.seed(1)
x <- sapply(1:10,function(x){paste(LETTERS[sample(1:26,10,replace = TRUE)],collapse="")})
y <- sapply(1:20,function(x){paste(LETTERS[sample(1:26,10,replace = TRUE)],collapse="")})
z <- rep(c("A","B"), times = 10)
df <- data.frame(x, y, z)
g <- ggplot(df, aes(x, y))+
  geom_point()+
  facet_wrap(~z, nrow =1) +
  xlab("X TITLE") +
  ylab("Y TITLE") +
  theme(axis.text.x = element_text(angle = 90))
g
gp <- ggplotly(g)
gp
gp \%>\%
  add_axis_title_spacing_plotly("x", 3) \%>\%
  add_axis_title_spacing_plotly("y", 3)

}
