% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{adjust_axis_title_spacing_plotly}
\alias{adjust_axis_title_spacing_plotly}
\title{Add Axis Spacing for Faceted Plotly}
\usage{
adjust_axis_title_spacing_plotly(plotly_obj, axis, adjustment)
}
\arguments{
\item{plotly_obj}{An object generated by ggplotly}

\item{axis}{String value of "x" or "y". Which axis title to add spacing to.
If need to add spacing to both, make separate calls to \code{adjust_axis_title_spacing_plotly()}}

\item{adjustment}{Numeric value. Vertical (for x-axis) or horizontal (for y-axis) adjustment for axis title}
}
\value{
A the original plotly object with updated axis name
}
\description{
Adjusts the axis titles in plotly objects by modifying the object itself,
since plotly's auto spacing does not do this appropriately for faceted plots
}
\examples{
library(ggplot2)
library(plotly)
library(dplyr)
set.seed(1)
x <- sapply(1:10,function(x){paste(LETTERS[sample(1:26,10,replace = TRUE)],collapse="")})
y <- sapply(1:20,function(x){paste(LETTERS[sample(1:26,10,replace = TRUE)],collapse="")})
z <- rep(c("A","B"), times = 10)
df <- data.frame(x, y, z)
g <- ggplot(df, aes(x, y))+
  geom_point()+
  facet_wrap(~z, nrow =1) +
  xlab("X TITLE") +
  ylab("Y TITLE") +
  theme(axis.text.x = element_text(angle = 90))
g
gp <- plotly::ggplotly(g)
gp
gp \%>\%
  adjust_axis_title_spacing_plotly("x", 0.05) \%>\%
  adjust_axis_title_spacing_plotly("y", 0.05)

}
