% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_formatting.R
\name{determine_passing_spec}
\alias{determine_passing_spec}
\title{Determine passing specification}
\usage{
determine_passing_spec(
  values,
  lower_threshold = NULL,
  upper_threshold = NULL,
  digits_round = NULL,
  pass_at_threshold = TRUE,
  flag_values = c(FALSE, TRUE)
)
}
\arguments{
\item{values}{Numeric vector of values to evaluate for spec}

\item{lower_threshold}{Numeric value that represents the lower limit of acceptable values}

\item{upper_threshold}{Numeric value that represents the upper limit of acceptable values}

\item{digits_round}{The number of digits that values are to be rounded to before evaluation}

\item{pass_at_threshold}{Logical value. Whether or not values equal to threshold should pass, defaults to TRUE.}

\item{flag_values}{Vector of length 2 where the first value is the value to be returned for passing/non-flagged values
and the second value is the value to be returned for non-passing/flagged values.}
}
\value{
Vector of length equal to input values. Output types defined by flag_values parameter
}
\description{
This function is currently unused. Confirm functionality and modify if needed before use.
Flags numeric values above and/or below supplied upper and lower thresholds.
}
\examples{
myvalues <- 1:20
flags <- determine_passing_spec(myvalues,
lower_threshold = 5,
upper_threshold = 18,
pass_at_threshold = TRUE)
flags
flags2 <- determine_passing_spec(myvalues,
lower_threshold = 5,
upper_threshold = 18,
pass_at_threshold = FALSE)
flags2
}
