% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get_cell_label_palette}
\alias{get_cell_label_palette}
\title{Load Cell Label Color Palette}
\usage{
get_cell_label_palette(cell_label_method = "Seurat3", cell_labels = NULL)
}
\arguments{
\item{cell_label_method}{Character value. Cell labeling method. Currently only
defined method is Seurat3.}

\item{cell_labels}{Character vector. All cell type labels in the dataset.}
}
\value{

}
\description{
For recognized labeling methods loads a cell-type specific color palette,
otherwise genreates a varibow palette based on number of cell types in data
}
\examples{
get_cell_label_palette(cell_label_method = "Seurat3")
\dontrun{get_cell_label_palette(cell_label_method = "SpecialLabelMethod") ## generates intentional error}
get_cell_label_palette(cell_label_method = "SpecialLabelMethod",
                       cell_labels = c("Label1","Label2","Label3"))
}
