% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_formatting.R
\name{gt_fmt_comments}
\alias{gt_fmt_comments}
\title{Format Flags in GT Table}
\usage{
gt_fmt_comments(x)
}
\arguments{
\item{x}{A gt table object. Must contain a column named "Comments" with flags "Warning" or "Fail".}
}
\value{
A gt table object with formatted Comments cells.
}
\description{
CURRENTLY NOT USED. Apply standard formatting to "Comments" column of a gt table
}
\details{
For "Comments" column, will fill cells containing pattern "Warning" in light red
and will fill cells containing pattern "Fail" in dark red.
}
\examples{
set.seed(1)
vComments <- sample(c("Pass","Warning","Fail",NA), 10, replace = TRUE)
my_gt <- gt::gt(data.frame(X = 1:10, Comments = vComments))
my_gt
gt_fmt_comments(my_gt)
my_gt_nocommments <- gt::gt(data.frame(X = 1:10, CommentsX = vComments))
my_gt_nocommments
\dontrun{gt_fmt_comments(my_gt_nocommments)  # This intentionally generates an error}
}
