% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{load_config}
\alias{load_config}
\title{Load config file}
\usage{
load_config(in_config, vector_delim = ";")
}
\arguments{
\item{in_config}{Character value. Full file path to the config file.}

\item{vector_delim}{Character value. Delimiter to use to convert string to vector}
}
\value{
A list of parameters. Names from "parameter" and values from "value"
with classes defined per "type"
}
\description{
Loads a csv config file of analysis parameters
}
\details{
Parameter file should have columns "parameter" containing parameter names,
"value" containing parameter values, and "type" containing variable classes.
}
\examples{
my_config <- system.file("extdata/default_rna_config_v1.csv", package = "batchreporter")
cnfg <- load_config(my_config)
}
