% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_umap_report}
\alias{plot_umap_report}
\title{Plot UMAP}
\usage{
plot_umap_report(
  df,
  x_col,
  x_lab,
  y_col,
  y_lab,
  title,
  point_size,
  color_col,
  scale_color_fun,
  ...
)
}
\arguments{
\item{df}{A data frame containing umap data and metadata for plotting}

\item{x_col}{Character value. Name of data column to plot on the x-axis}

\item{x_lab}{Character value. Name to display on the x-axis}

\item{y_col}{CHaracter value. Name of data column to plot on the y-axis}

\item{y_lab}{Character value. Name to display on the y-axis}

\item{title}{Character value. Title of plot.}

\item{point_size}{Numeric value. Point size to use for scatter plot}

\item{color_col}{Character value, Name of data column used to color point by groups or value}

\item{scale_color_fun}{A function that returns color scaling for ggplot2}

\item{...}{Additional arguments passed to \code{scale_color_fun}}
}
\value{
A ggplot2 plot object
}
\description{
Standardized UMAP plotting with color customization for grouping variables
}
