% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{qc_piechart}
\alias{qc_piechart}
\title{QC Pie Chart with Optional Faceting}
\usage{
qc_piechart(
  meta,
  category_y,
  category_name = "Cell Type",
  stat = "count",
  variable_y_identity = NULL,
  palette_df_group = NULL,
  colorset_y = "varibow",
  plot_font_size = 14,
  label_size = 3,
  text_repel_x = 1.5,
  nudge_x_label = 0.5,
  facet_formula = NULL,
  ...
)
}
\arguments{
\item{meta}{A data.frame containing metadata}

\item{category_y}{A character value specifying the metadata column to use for color groups}

\item{category_name}{A character value specifying a name to display for the category colors}

\item{stat}{A character value, either "count" or "identity", depending on whether input data
has already been aggregated by the category_y grouping variable. Default is "count".}

\item{variable_y_identity}{Value to use as labels for the pie wedge. Currently supported
values are "p}

\item{palette_df_group}{An optional data frame containing two columns defining the
categorical data color palette. The first column contains label values for
all levels of category_y and the second column contains the color for each
label. If NULL (default) random colors will be generated via the colorset_y
parameter.}

\item{colorset_y}{A colorset to use as fills for category_y. Currently
supported: "rainbow" or "varibow". Default is "varibow". Only used when
palette_df_group is NULL.}

\item{plot_font_size}{Numeric value. General size of text on plot in font size. Default is 14.}

\item{label_size}{Size of text to label each pie wedge. Default is 3}

\item{text_repel_x}{Numeric value. X coordinate for plotting pie wedge labels
using \code{ggrepel::geom_text_repel()}. Default is 1.5}

\item{nudge_x_label}{Numeric value. Distance to nudge label in X coordinate
for plotting pie wedge labels using \code{ggrepel::geom_text_repel()}. Higher
values will increase line length. Default is 0.5.}

\item{facet_formula}{A formula object for faceting based on variables in the meta data frame.
For example, \code{formula("~pool_id")} will facet wrap by a variable called pool_id in meta.
Currently only actual column names (rather than variables) supported for this argument.}

\item{...}{Additional arguments passed to \code{ggplot2::facet_wrap()}}
}
\value{
A ggplot2 plot object
}
\description{
Create a pie chart with wedges specified by a grouping variable. Faceting may be performed
if facetting is provided.
}
\details{
The grouping variable used for category_y will generate the wedges of the bar plot. The supplied
dataframe can either be one observation per row or one row per category_y. If the former, "stat" should be
"count", in which case counts of each category will be calculated. If the latter, "stat" should be
"identity", and a column containing the category counts should be supplied ("variable_y_identity").
Facet wrapping will be performed on supplied variables based on the facet_formula.
}
\examples{
\dontrun{
set.seed(2021)
my_palette <- batchreporter::get_cell_label_palette()
cell_probs <- c(0.025, 0.20, 0.03, 0.192, 0.18, 0.14, 0.05,
                0.005, 0.02, 0.1 ,0.001,0.008,0.049)
test_data_1 <- data.frame(cell_type = sample(my_palette[,1], 500, 
                                             prob = cell_probs, replace = TRUE),
                        group = sample(c("MT UMI>0.1", "MT UMI<0.1"), 500, 
                                       prob = c(0.1, 0.9), replace = TRUE),
                        sample_names= 1:500)
# Use a predefined palette
qc_piechart(test_data_1,
            category_y = "cell_type", 
            category_name = "Cell Type",
            stat = "count",
            palette_df_group = my_palette,
            plot_font_size = 14,
            label_size = 3,
            text_repel_x = 1.5, 
            nudge_x_label = 0.2,
            facet_formula = as.formula("~group"))
# Use default random palette
qc_piechart(test_data_1,
            category_y = "cell_type",
            stat = "count",
            label_size = 3,
            text_repel_x = 1.5, 
            nudge_x_label = 0.2,
            facet_formula = as.formula("~group"))

# No Facetting
qc_piechart(test_data_1,
            category_y = "cell_type",
            stat = "count",
            label_size = 3,
            text_repel_x = 1.5, 
            nudge_x_label = 0.2)

# Stat = "identity" version, precalculated counts
test_data_2 <- data.frame(cell_type = rep(my_palette[,1], times = 2),
                         group = rep(c("MT UMI>0.1", "MT UMI<0.1"), each = nrow(my_palette)),
                         sample_names= 1:(2*nrow(my_palette)),
                         group_count = c(round(1000*cell_probs),round(4000*sample(cell_probs))))
test_data_2
qc_piechart(test_data_2,
            category_y = "cell_type",
            stat = "identity",
            variable_y_identity = "group_count",
            label_size = 3,
            text_repel_x = 1.5, 
            nudge_x_label = 0.2,
            facet_formula = as.formula("~group"))
} # end don't run
}
