% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{qc_stacked_barplot_facet}
\alias{qc_stacked_barplot_facet}
\title{QC Stacked Barplot with Faceting}
\usage{
qc_stacked_barplot_facet(
  meta,
  category_x = "batch_id",
  name_x = "Batch ID",
  category_y = "well_id",
  category_name = "Well ID",
  colorset_y = "varibow",
  name_y = "N Cells",
  as_fraction = FALSE,
  facet_formula = NULL,
  ...
)
}
\arguments{
\item{meta}{A data.frame containing metadata}

\item{category_x}{A character object specifying the metadata to use for splitting in the x-direction}

\item{name_x}{A character object specifying a name to display on the x-axis}

\item{category_y}{A character object specifying the metadata to use for color groups}

\item{category_name}{A character object specifying a name to display for the colors}

\item{colorset_y}{A colorset to use as fills for category_y. Currently supported: "rainbow" or "varibow". Default is "varibow"}

\item{name_y}{A character object specifying a name for the y-axis.}

\item{as_fraction}{A logical object specifying whether or not to display the stacked bars as fractions of the total count for each category_x. Default is FALSE.}

\item{facet_formula}{A formula object for faceting based on variables in the meta data frame. For example, \code{formula("~pool_id")} will facet wrap by a variable called pool_id in meta.}

\item{...}{Additional arguments passed to \code{ggplot2::face_wrap()}}
}
\value{
A ggplot2 plot object
}
\description{
The metric used for category_x will generate bars as columns on the x-axis. The bar will be split vertically based on category_y.
Facet wrapping will be performed on supplied variables based on the facet_formula.
}
