% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{read_h5_well_meta}
\alias{read_h5_well_meta}
\title{Read H5 Well Metadata}
\usage{
read_h5_well_meta(h5_file, target = "well")
}
\arguments{
\item{h5_file}{the path to an .h5 file with metadata injection by AIFI pipeline.}

\item{target}{A specific matrix object in the .h5 file. Default is "well". Can also work with any matrix-like object
nested further if target is equal to the full path (ie matrix/features)--matrix will be read in as-is.
For reading cell level metadata use \code{H5weaver::read_h5_cell_metadata()} and for reading
sparse matrixes use \code{H5weaver::read_h5_dgCMatrix()}}
}
\value{
A data.frame containing all feature metadata found in /well
}
\description{
Read in 10x well metadata from an H5 file. Metadata is injected to the
10x Genomics cellranger h5 output by AIFI pipeline code.
}
