% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{read_hto_well_json}
\alias{read_hto_well_json}
\title{Read in HTO Processing JSON Files}
\usage{
read_hto_well_json(fp)
}
\arguments{
\item{fp}{Full file path to a single json file to be read. JSON files are generated
by the AIFI pipeline}
}
\value{
A data frame of select hashing information for each well and HTO
}
\description{
Read in HTO Processing JSON Files
}
