% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna_analysis.R
\name{sample_cells}
\alias{sample_cells}
\title{Sample cells}
\usage{
sample_cells(cl, sample.size, seed = 3030)
}
\arguments{
\item{cl}{Named character vector. Values are cell labels and names are cell barcodes.}

\item{sample.size}{Integer value. Number of cells of each cell type to sample}

\item{seed}{Integer value. Seed passed to \code{set.seed()} for reproducible sampling.}
}
\value{
Named character vector. Values are the cell barcodes selected and names are the labels
}
\description{
Samples cells on labels, retaining an equal number of each label type.
}
