% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{scale_color_fct_mito}
\alias{scale_color_fct_mito}
\title{Default Color Gradient for Fraction Values}
\usage{
scale_color_fct_mito(
  colors = c("blue", "green3", "yellow", "red"),
  breaks = c(0, 0.25, 0.5, 0.75, 1),
  ...
)
}
\arguments{
\item{colors}{Colors for \code{scale_color_gradientn()}.}

\item{breaks}{Breaks for \code{scale_color_gradientn()}.}

\item{...}{Additional arguments passed to \code{scale_color_gradientn()}}
}
\value{
A custom function of \code{scale_color_gradientn()} for fractional values with colors, values,
and breaks defined. The function will accept additional arguments to \code{scale_color_gradientn()}
}
\description{
Creates a default color gradient for fractional values using \code{scale_color_gradientn()}
}
