% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{scale_color_genes}
\alias{scale_color_genes}
\title{Default Color Gradient for Genes per Cell}
\usage{
scale_color_genes(max_value)
}
\arguments{
\item{max_value}{Character value. Highest genes per cell value in the dataset}
}
\value{
A custom function of \code{scale_color_gradientn()} with colors, values,
and breaks defined. The function will accept additional arguments to \code{scale_color_gradientn()}
}
\description{
Uses \code{} with default values to create a gradient for genes per cell, with
scale adjusting to observed values
}
