% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{seurat_3_cell_palette}
\alias{seurat_3_cell_palette}
\title{Seurat 3 Cell Palette}
\usage{
seurat_3_cell_palette()
}
\value{
A data frame with columns "seurat3_pbmc_type" and "cell_color"
}
\description{
Color palette with color assigned to each cell type in Seurat v3 labeling
reference (variable_pbmc_10k_v3)
}
