% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_formatting.R
\name{simple_html_table}
\alias{simple_html_table}
\title{Create a simple HTML table of key and value pairs}
\usage{
simple_html_table(labels, values, col_widths_px = c(300, 800), fontsize = 2)
}
\arguments{
\item{labels}{Vector of labels}

\item{values}{Vector of values, same length as labels}

\item{col_widths_px}{Numeric vector of length 2. Widths in pixels of each column,}

\item{fontsize}{Numeric value. Default 2.}
}
\value{
HTML of formatted table.
}
\description{
Create a simple table of paired values (ie name and value).
}
\examples{
lab1 <- c("Batch", "N Samples", "N Pools")
val1 <- c("B000", 16, 2)
h_tab <- simple_html_table(lab1, val1)
htmltools::html_print(h_tab)
}
