% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{qc_aligned_barplot_facet}
\alias{qc_aligned_barplot_facet}
\title{QC Aligned Barplot with Faceting}
\usage{
qc_aligned_barplot_facet(
  meta,
  category_x = "batch_id",
  name_x = "Batch ID",
  category_y = "well_id",
  category_name = "Well ID",
  name_y = "N Cells",
  padding = 0.2,
  colorset_y = "varibow",
  stat = "count",
  variable_y_identity = NULL,
  facet_formula = NULL,
  ...
)
}
\arguments{
\item{meta}{A data.frame containing metadata}

\item{category_x}{A character object specifying the metadata to use for splitting in the y-direction}

\item{name_x}{A character object specifying a name to display on the x-axis}

\item{category_y}{A character object specifying the metadata to use for color groups}

\item{category_name}{A character object specifying a name to display for the colors}

\item{name_y}{A character object specifying a name for the y-axis.}

\item{padding}{A numeric object specifying the the fraction of the total vertical space to use for separating category_y groups. Default is 0.2.}

\item{colorset_y}{A colorset to use as fills for category_y. Currently supported: "rainbow" or "varibow". Default is "varibow"}

\item{facet_formula}{A formula object for faceting based on variables in the meta data frame. For example, \code{formula("~pool_id")} will facet wrap by a variable called pool_id in meta.}

\item{...}{Additional arguments passed to \code{ggplot2::face_wrap()}}
}
\value{
A ggplot2 plot object
}
\description{
The metric used for category_x will generate bars as columns on the x-axis. The bar will be split vertically based on category_y. Each group in category_y will be aligned to make them easier to compare.
Facet wrapping will be performed on supplied variables based on the facet_formula.
}
